test: move all unit tests in their own dir
