+++ tests/unit/test-firefox/test.firefox.js
@@ -6,14 +6,14 @@ import deepcopy from 'deepcopy';
 import sinon from 'sinon';
 import FirefoxProfile from 'firefox-profile';
 
-import * as firefox from '../../src/firefox';
-import {onlyInstancesOf, WebExtError} from '../../src/errors';
+import * as firefox from '../../../src/firefox';
+import {onlyInstancesOf, WebExtError} from '../../../src/errors';
 import {fs} from 'mz';
-import {withTempDir} from '../../src/util/temp-dir';
+import {withTempDir} from '../../../src/util/temp-dir';
 import {TCPConnectError, fixturePath, fake, makeSureItFails} from '../helpers';
 import {basicManifest, manifestWithoutApps} from '../test-util/test.manifest';
-import {defaultFirefoxEnv} from '../../src/firefox/';
-import {RemoteFirefox} from '../../src/firefox/remote';
+import {defaultFirefoxEnv} from '../../../src/firefox/';
+import {RemoteFirefox} from '../../../src/firefox/remote';
 
 
 describe('firefox', () => {
