+++ tests/unit/test-util/test.logger.js
@@ -5,7 +5,7 @@ import bunyan from 'bunyan';
 import sinon from 'sinon';
 import {WriteStream} from 'tty';
 
-import {createLogger, ConsoleStream} from '../../src/util/logger';
+import {createLogger, ConsoleStream} from '../../../src/util/logger';
 
 
 describe('logger', () => {
