+++ tests/unit/test.errors.js
@@ -2,7 +2,7 @@
 import {describe, it} from 'mocha';
 import {assert} from 'chai';
 
-import {onlyErrorsWithCode, onlyInstancesOf} from '../src/errors';
+import {onlyErrorsWithCode, onlyInstancesOf} from '../../src/errors';
 import {makeSureItFails} from './helpers';
