+++ tests/unit/test.setup.js
@@ -1,6 +1,6 @@
 /* @flow */
 import {beforeEach, afterEach} from 'mocha';
-import {consoleStream} from '../src/util/logger';
+import {consoleStream} from '../../src/util/logger';
 
 beforeEach(function() {
   consoleStream.makeVerbose();
